lib: move bit buffer to toolbox
