+++ lib/nfc/helpers/iso14443_crc.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-#include "bit_buffer.h"
+#include <toolbox/bit_buffer.h>
 
 #ifdef __cplusplus
 extern "C" {
