+++ lib/nfc/protocols/iso14443_3a/iso14443_3a_poller_i.h
@@ -2,7 +2,7 @@
 
 #include "iso14443_3a_poller.h"
 
-#include <nfc/helpers/bit_buffer.h>
+#include <toolbox/bit_buffer.h>
 
 #ifdef __cplusplus
 extern "C" {
