+++ lib/nfc/protocols/iso14443_3b/iso14443_3b.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <lib/nfc/helpers/bit_buffer.h>
+#include <toolbox/bit_buffer.h>
 #include <nfc/protocols/nfc_device_base_i.h>
 
 #ifdef __cplusplus
