+++ lib/nfc/protocols/iso15693_3/iso15693_3_poller_i.h
@@ -4,7 +4,7 @@
 
 #include "iso15693_3_i.h"
 
-#include <nfc/helpers/bit_buffer.h>
+#include <toolbox/bit_buffer.h>
 
 #ifdef __cplusplus
 extern "C" {
