+++ lib/nfc/protocols/mf_classic/crypto1.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <lib/nfc/helpers/bit_buffer.h>
+#include <toolbox/bit_buffer.h>
 
 #ifdef __cplusplus
 extern "C" {
