+++ lib/toolbox/bit_buffer.c
